+++ R/moran.R
@@ -1,17 +1,25 @@
 #' Computes Moran's I correlation index
 #'
 #' Natively built for computing Moran's I on \code{dgCMatrix} objects, this
-#' routine allows computing the I on large sparse matrices (graphs), feature that
-#' is not supported on \code{\link[ape:Moran.I]{ape::Moran.I}}.
+#' routine allows computing the I on large sparse matrices (graphs). Part of
+#' its implementation was based on \code{\link[ape:Moran.I]{ape::Moran.I}},
+#' which computes the I for dense matrices.
 #'
 #' @param x Numeric vector of size \eqn{n}.
 #' @param w Numeric matrix of size \eqn{n\times n}{n * n}. Weights. It can be
 #'  either a object of class \code{\link{matrix}} or \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}
 #'  from the \code{\link[Matrix]{Matrix}} package.
 #' @param normalize.w Logical scalar. When TRUE normalizes rowsums to one (or zero).
+#' @param alternative	Character String. Specifies the alternative hypothesis that
+#' is tested against the null of no autocorrelation; must be of one \code{"two.sided"},
+#' \code{"less"}, or \code{"greater"}.
 #' @export
 #' @family statistics
-#' @return Numeric scalar with Moran's I.
+#' @return A list of class \code{diffnet_moran} with the following elements:
+#' \item{observed}{Numeric scalar. Observed correlation index.}
+#' \item{expected}{Numeric scalar. Expected correlation index equal to \eqn{-1/(N-1)}.}
+#' \item{sd}{Numeric scalar. Standard error under the null.}
+#' \item{p.value}{Numeric scalar. p-value of the specified \code{alternative}.}
 #' @references
 #' Moran's I. (2015, September 3). In Wikipedia, The Free Encyclopedia.
 #' Retrieved 06:23, December 22, 2015, from
@@ -23,7 +31,7 @@
 #' # Generating a small random graph
 #' set.seed(123)
 #' graph <- rgraph_ba(t = 4)
-#' w <- igraph::distances(igraph::graph_from_adjacency_matrix(graph))
+#' w <- approx_geodesic(graph)
 #' x <- rnorm(5)
 #'
 #' # Computing Moran's I
@@ -34,7 +42,7 @@
 #' ape::Moran.I(x, w)
 #' }
 #' @author George G. Vega Yon
-moran <- function(x, w, normalize.w=TRUE) {
+moran <- function(x, w, normalize.w=TRUE, alternative = "two.sided") {
   if (!inherits(w, "matrix") & !inherits(w, "dgCMatrix"))
     stop("-w- must be either a matrix or a dgCMatrix.")
 
@@ -45,10 +53,30 @@ moran <- function(x, w, normalize.w=TRUE) {
     w <- methods::as(w, "dgCMatrix")
 
   if (normalize.w)
-    w <- w/(rowSums(w) + 1e-15)
+    w <- w/(rowSums(w) + 1e-20)
 
   res <- moran_cpp(x, w)
-  names(res) <- c("I", "Var")
-  res
+
+  # Computing pval
+  pv  <- with(res, pnorm(observed, mean = expected, sd = sd))
+
+  # Checking alternatives
+  alternative <- match.arg(
+    alternative, c("two.sided", "less", "greater")
+    )
+
+  if (alternative == "two.sided")
+    pv <- if (res$observed <= res$expected)
+      2 * pv
+  else 2 * (1 - pv)
+  if (alternative == "greater")
+    pv <- 1 - pv
+
+  # Returning
+  structure(
+    c(res, p.value = pv),
+    class = "diffnet_moran"
+  )
+
 }