+++ src/stats.cpp
@@ -132,7 +132,7 @@ arma::sp_mat vertex_covariate_compare(const arma::sp_mat & graph, const NumericV
 }
 
 // [[Rcpp::export]]
-std::vector<double> moran_cpp(const arma::colvec & x, const arma::sp_mat & w) {
+List moran_cpp(const arma::colvec & x, const arma::sp_mat & w) {
   double xmean = mean(x);
 
   int N = x.n_rows;
@@ -148,40 +148,46 @@ std::vector<double> moran_cpp(const arma::colvec & x, const arma::sp_mat & w) {
   double numer = accu((xcent * xcent.t()) % w);
   double denom = accu(pow(xcent, 2.0));
 
-  /*/ Computing standard error
-  double s1, s2, s3, s4, s5;
-  arma::sp_mat::const_iterator b = w.begin();
-  arma::sp_mat::const_iterator e = w.end();
+  // Computing standard error
+  double s1 = 0.0, s2 = 0.0, s3, s4, s5;
+  arma::sp_mat::const_iterator b;
 
-  int i,j;
-  for (arma::sp_mat::const_iterator iter=b; iter!=e; ++iter) {
-  i = iter.col();
-  j = iter.row();
+  // S1
+  unsigned int i,j;
+  arma::sp_mat wtw = w + w.t();
+  for (b = wtw.begin(); b!= wtw.end(); b++)
+    s1 += pow((*b), 2.0);
 
-  // Computing each count
-  if (i >= j) s1+= pow(w.at(i,j) + w.at(j,i),2.0);
+  s1/= 2.0;
 
-  }
-  s1 = sqrt(s1);
+  // S2
 
-  arma::colvec tmp(w.n_cols);
-  for (int i=0; i < (int) w.n_rows; i++)
-  tmp.at(i) = pow(accu(w.row(i)) + accu(w.col(i)), 2.0);
+  // row sums
+  arma::sp_colvec w_rowsums = sum(w, 1);
+  arma::sp_rowvec w_colsums = sum(w, 0);
 
-  s2 = sum(tmp);
+  for (i = 0; i < N; i++)
+    s2 += pow(w_rowsums.at(i) + w_colsums.at(i), 2.0);
 
-  s3 = (1.0/N * accu(pow(xcent,4.0))) / pow(1.0/N * accu(pow(xcent,2.0)), 2.0);
-  s4 = (N*N - 3.0*N + 3.0)*s1 - N*s2 + 3.0 * wsum * wsum;
-  s5 = (N*N - N)*s1 - 2.0*N*s2 + 6.0*wsum*wsum;
-  */
-  std::vector<double> res(2);
-  res[0] = (x.size()/wsum)*(numer/(denom + 1e-15));
+  // S3
+  s3 = accu(pow(xcent, 4.0))/N/
+    pow(accu(pow(xcent, 2.0))/N, 2.0);
 
-  /*res[1] = (N*s4 - s3*s5)/((N - 1.0)*(N - 2.0)*(N - 3.0)*wsum*wsum) -
-   pow(-1.0/(N - 1.0), 2.0);*/
-  res[1] = 0;
+  // S4
+  s4 = (N*N - 3.0 * N + 3.0)*s1 - N*s2 + 3*pow(wsum, 2.0);
 
-  return res;
+  // S5
+  s5 = N*(N - 1.0)*s1 - 2.0*N*s2 + 6.0*pow(wsum, 2.0);
+
+  return Rcpp::List::create(
+    _["observed"] = (N/wsum)*(numer/(denom + 1e-20)),
+    _["expected"] = -1.0 / (N - 1.0),
+    _["sd"] = sqrt(
+      (N*s4 - s3*s5)/((N - 1.0)*(N - 2.0)*(N - 3.0) * pow(wsum, 2.0)) -
+      pow(-1.0/(N - 1.0), 2.0)
+    )
+  );
+
 }